Fix error message when attestation is disabled
When attestation rate limiting is disabled in the middleware, the
AttestAgent RPC is still invoking the rate limiter. This causes the
middleware to log that there is a bug, since an RPC was either not
limited (and not expected to invoke the limiter) or was limited (and
expected to invoke the limiter).

This change introduces a new kind of limit, the "disabled" limit and
fixes the middleware layer to still expect the rate limiter to be
invoked by the RPC for disabled limits.

It then changes the rate limiting configuration to use the "disabled"
limit for Agent.AttestAgent when attestation rate limiting is disabled.

Signed-off-by: Andrew Harding <aharding@vmware.com>