Introduce Agent API middleware
This change introduces a middleware layer into the agent, which
simplifies some of the boilerplate present in the API handlers.

Specifically, the middleware:
- Handles security header enforcement
- Embellishes loggers with caller information (RPC being called, PID of
caller, etc).
- Records telemetry for the RPC call and maintains global per-service
connection counts
- Consolidates peertracker interaction to a single layer

Beyond the addition of the middleware, it provides:
- Exhaustive unit-testing of the endpoints package
- Cleans up some unused configuration for the endpoints package
- Moves the middleware infrastructure to a common package and provides
aliases where appropriate (so the server implementation doesn't have to
import multiple packages with the same name)

Signed-off-by: Andrew Harding <aharding@vmware.com>