Fix race in "spire" UpstreamAuthority plugin
The "spire" UpstreamAuthority plugin maintains a local copy of the
upstream bundle. When the upstream bundle changes, updates are sent back
to SPIRE core. The local copy is adjusted synchronously with the
response of NewDownstreamX509CA and PublishJWTAuthority RPCs. However,
there is also an asynchronous goroutine polling for updates to the bundle.
To prevent the pollling goroutine from overwriting the results of the
aforementioned RPCs, the plugin maintains a simple "version" number that
is bumped whenever the bundle is updated. The polling goroutine captures
the version number before polling for the bundle and is supposed to only
update the local copy if the local copy hasn't otherwise been updated
(by NewDownstreamX509CA or PublishJWTAuthority responses).

However, the version number check and update operation, i.e.
"replace if version matches" operation does not take place under the
same lock, opening up a race condition where the version check succeeds,
the local copy is updated by other goroutines, and then those updates
are overwritten by the now-stale bundle retrieved by the polling
goroutine.

This PR fixes that race condition by performing the version check and
replacement while under the lock.

It also fixes races in the unit-tests due to the wrong mock clock
implemtnation being used and not triggering timers enough to ensure
updates have been picked up before asserting the response.

Signed-off-by: Andrew Harding <aharding@vmware.com>