Fork SDS implementation into v2 and v3
We need to continue supporting older versions of Envoy, but support for
v2 will be dropping in Q1 2020. So for now, fork the handler into v2 and
v3 versions. At the point where we no longer support the last version of
Envoy to support v2 (v1.16.0?), we can cull the v2 implementation.

Signed-off-by: Andrew Harding <aharding@vmware.com>