Stop logging JWT-SVIDs that fail validation
While reworking the workload handler unit-tests, I noticed that we are
logging JWT-SVIDs that fail validation. This is a security concern since
it means that those who are able to inspect the agent log can get their
hands on JWT-SVIDs, which might otherwise be valid even though the agent
rejected them (e.g. mismatched audience values, workload not federated
with the JWT-SVID trust domain).

Looks like this was added by #1059 while expanding the logging around
telemetry emission.

This change removes the logging of the SVID.

Signed-off-by: Andrew Harding <aharding@vmware.com>