Move entry cache hydration out of request context
The in-memory entry and Agent selector cache is currently re-built
within the context of a FetchAuthorizedEntries() RPC request. In some
environments, re-hydrating the cache may be an expensive operation. If
the cache takes a long time to rebuild, it can degrade availability of
the FetchAuthorizedEntries() RPC for Agents because Agents may time out
while the cache is rebuilding. If the cache cannot be rebuilt within
the Agent's client timeout period, the Server may become completely
unavailable, as each Agent request times out waiting for the cache to
rebuild.

Building the cache within the FetchAuthorizedEntries() request context
also has the disadvantage of skewing p99 RPC latencies. Requests
that have to rebuild the cache or wait on another request to rebuild the
cache may take exponentially longer to complete than p50 requests. This
makes it hard to deterministically estimate the latency of p99 requests
and also adds more complexity to understanding the root cause of these
elevated latencies.

For these reasons, moving the hydration of the entry cache into a
dedicated goroutine that runs on a fixed interval (currently set to 5
seconds, the Agent's sync interval). Agents can continue to fetch the
latest entries from the cache while the cache is being re-hydrated.

Signed-off-by: Ryan Turner <turner@uber.com>