Close plugins that fail to configure
There is a bug in the catalog that prevents a plugin from being closed if it
encounters a failure to configure. Even though spire-server shuts down, this
can cause external plugin processes to be orphaned.

This is a little tricky to test via unit-testing but I did do manual
verification. However, the catalog tests where orphaning the external test
plugin process during the course of the testing. This change fixes that issue.

Signed-off-by: Andrew Harding <aharding@vmware.com>