Fix CLI exit code on entry create/update failure
Recent changes to adopt the batch entry create/update APIs broke the
exit code when the operation fails.

This change fixes the exit code.

It also updates the CLI framework to prefix errors with "Error: " to
maintain casing consistency with the other messages printed by the CLI.

Also, the CLI output when entry creation/updating fails is a little
cryptic.  The status message in particulary is printed after the
"failed" entry following a newline, which causes it to be grouped with
the next failed entry (when there is more than one entry that failed).
This change updates the output so that the status code and message are
grouped clearly with the offending entry. It also prints the text to
stderr instead of stdout.

Signed-off-by: Andrew Harding <aharding@vmware.com>