Migrate OIDC Discovery Provider to new APIs
This change updates the OIDC Discovery Provider to use the new server
APIs (namely the Bundle API) instead of the deprecated Registration API.

The `registration_api` configuration section is deprecated in favor of a
new `server_api` configuration section. The only difference between the
two sections is that instead of a socket path configurable, the
`server_api` configuration section has an address configurable, which is
a gRPC target address. Only the unix name system is supported.

The change to `address` is forward thinking to a future point where we
might allow the bundle to be fetched over the TCP listener of the
server, since the Bundle.GetBundle RPC is unauthenticated. I'm punting
on adding that support now to keep this change scoped to simply moving
away from the deprecated APIs (adding support now would neccesitate code
to bootstrap the provider with a bundle to authenticate the server
X509-SVID).

Signed-off-by: Andrew Harding <aharding@vmware.com>