Remove SPIFFE ID label from metrics
There were three remaining counter metrics in the Server CA component
that were including a SPIFFE ID as a metric label. A SPIFFE ID is too
high of cardinality to be used as a metric label because it can cause
indexing problems in metrics backends.

In all three cases, the SPIFFE ID is already being logged in the same
place where the metric is emitted, so this information is still
available through log entries.

Signed-off-by: Ryan Turner <turner@uber.com>