Update go-sqlite3 to v1.14.6
Not only is our current dependent version over a year and a half old, the updated
amalgamated code contains a fix for an annoying warning for those of us
compiling with with GCC 10, which shows up when building/testing SPIRE
(i.e. anything that transitively depends on the sql plugin, which includes
all tests).

Signed-off-by: Andrew Harding <aharding@vmware.com>