Fix file descriptor leak in linux peertracker
Fixes #2042

statfd was never closed, so there was a file descriptor leak that
eventually got cleaned up on garbage collection due to a finalizer
on os.File.

Since the file is only opened to read the contents, the current logic
can be simplified by using `ioutil.ReadFile` which opens the file, reads
the contents, and closes the file.

Signed-off-by: Prashant Varanasi <prashant@uber.com>