Allow server compilation w/o CGO
SPIRE server has a dependency on CGO brought in with the support of
sqlite3. Currently you can't build the server, since when CGO is
disabled, the mattn/go-sqlite3 package is more or less empty, and some
types we use are not available, causing the build to fail.

This change provides support for compiling the server with CGO disabled
by leveraging build tags to only pull in the sqlite3 support when CGO is
enabled, and provide a small shim that returns a specific error on
sqlite3 connection attempts when not.

Signed-off-by: Andrew Harding <aharding@vmware.com>