Declare supported Envoy versions
This change introduces documentation on the supported Envoy versions. It
also embellishes the Envoy integration test (for v3) to test against
each supported minor version.

The integration test relies on the vX.Y-latest tags documented in the
official Docker Hub repository for envoy, which stats that the tags with
the latest suffix point to the latest point release for that minor
version.

Fixes: #1499

Signed-off-by: Andrew Harding <aharding@vmware.com>