Fix upgrade integration test version check
The upgrade integration test does some sanity checking to ensure that
the code under test reflects the proper version. This helps us remember
to bump the in-code version after making a release.

Unfortunately, the version check is broken and does not successfully
detect the prior release version for the PR target branch due both to
the use of git describe and various scripting bugs. This change
fixes the test to properly detect the most recent version of the branch
(or of the project in the case of master). It relies on the CI/CD
pipeline setting the proper target branch for the PR in the environment.

Additionally, Travis only clones to depth 50. If there are more than 50
commits since the last release, then the test fails to detect the prior
version. This change bumps the clone depth to 100. It is a bit of a
band-aid but still serves to reduce the amount of data needing to be
cloned. If we surpass this limit, it might make sense to just clone the
whole repository. At the time of this change, this is the difference
between 2MB and 12MB downloaded, but that delta would increase with
time.

Signed-off-by: Andrew Harding <aharding@vmware.com>