Fix upgrade integration test for merge commits
On CircleCI, the newly introduced version check logic fails on merge
commits because there is no PR number from which to determine the target
branch.

This change fixes it by providing a small script that returns the target
branch, following the same semantics as the TRAVIS_BRANCH environment
variable used in the Travis CI/CD pipeline.

It also renames the PR_BRANCH variable to CICD_TARGET_BRANCH to more
accurately represent the value.

Signed-off-by: Andrew Harding <aharding@vmware.com>