Remove deprecated Node API
The Node API was officially deprecated in v0.12.0. However, the SPIRE
agent has not consumed it for since before 0.11.0. The handler code is
some of the hardest code to follow, maintain, and test. It is also a
"private" API, and was never designed for public consumption.
Considering these points and in preparation for v1.0.0, this change
removes the Node API.

This change does NOT impact the Registration API, which is public and
still in use by many organizations and will continue to be present for
the v1.0.x release cycle. However, it will likely be deprecated with the
v1.1.0 release.

Signed-off-by: Andrew Harding <aharding@vmware.com>