Fix race conditions in pkg/agent/svid
The TestRunWithUpdates test case has two race conditions which
cause intermittent test failures. First, the return value of the
Run method of the rotator is non-deterministic in the face of
context cancellations. Second, it is possible for the RenewSVID
function of the test mock to be called multiple times depending on
test timing conditions.

The first race is fixed by adjusting the test assertion to accept
either possible return value. The second race is fixed by removing
an extranous explicit call to rotateSVID

Signed-off-by: Brad Blackard <bradb@uber.com>