Fix excessive execution time in update package test
The test in the uptime package currently takes 60 seconds to run. This
is due to the call to reportMetrics blocking until the parent context
times out. That timeout was hard-coded to one minute.

This commit improves the execution time of the test by removing the custom
timeout and instead relying on explicit context cancellation when the
expected assertion conditions have been met. Additional it simplies the
test case logic.

Signed-off-by: Brad Blackard <bradb@uber.com>