JWT SVID exp validation
SPIFFE mandates that the "exp" claim of JWT SVIDs MUST NOT be a time in
the past. This is not explicitly validated in the JWT SVID CA component
of SPIRE Server. If the JWT signing key was not rotated properly, it
could result in a case where SPIRE Server issued JWT SVIDs with "exp"
claims in the past.

Explicitly validate this case rather than rely on proper Server
configuration validation and successful JWT key rotation.

Signed-off-by: Ryan Turner <turner@uber.com>