Remove dead code in server upstream client
The upstream_bundle has was removed in 0.11.0 but some related code
remained the upstream client, which would shut down the open upstream
bundle streams when upstream_bundle was not set.

This code is no longer meaningful, since the condition that triggers it
is not possible.

This change removes the dead code.

Signed-off-by: Andrew Harding <aharding@vmware.com>