Provide non-global test keys
The testkey package currently only provides a set of global helpers to
obtain test keys. This means that the number of keys a given test
package uses is an aggregate of all tests in that package, which makes
it really easy to exceed the number of pregenerated keys when you add
some additional test cases.

This small refactor allows tests to use a testkey Keys instance to have
per-test pregenerated keys, making it much less likely to exhaused the
pregenerated key pool.

Signed-off-by: Andrew Harding <aharding@vmware.com>