Fix broken rotator tests
The rotator tests try and assert rotator behavior when run with both a
good and expired SVID. However, the tests are flaky since they do not
reliably wait until the rotation check finishes.

Furthermore, the rotation test for an existing SVID is loading an
expired certificate from the test fixtures. This is against the
preconditions required for the test case, which are that it does not
rotate when a good SVID already exists.

These tests are super hacky. I'd clean them up now but I have hopes that
they don't survive in the short to mid term. I have a local branch that
I need to wrap up that cleans up the interactions between the attestor,
rotator, client, and manager packages, and replaces these tests as part
of the work to re-attest without shutting down the agent. One day I'll
get the time to finish up that work.

Signed-off-by: Andrew Harding <aharding@vmware.com>