New catalog with version support
- Introduces the idea of plugin and service repositories which advertise
their supported versions and available builtins and provide a binder for
the catalog to populate the repository with applicable facades.
- Generates code for the new SDK interfaces for existing
plugins/services. Removes the old spire-plugingen targets and tooling.
- Removes the reflect-heavy "fill" code with something a little more
explicit. The only use of reflect is to validate that the binder can
accept all version facades and wrap it for internal use.
- Removes examples and plugin test tools. Conceptually these belong in
the plugin SDK repository.
- Removes spiretest.LoadPlugin in favor of a new plugintest.Load system
tailored to the new catalog. The plugintest package makes for a clean
home for plugin test related things.
- `make generate` is now very fast, and has always been idempotent, so
it just regenerates every time it is called and doesn't try and be smart
about only regenerating what is needed. The plugingen and protogen
targets have been removed accordingly. This sidesteps problems with tool
version change detection that we've had historically.
- A few unit-tests were converted to test through the facade interface.
The rest were left until the individual plugin is converted to the v1
interface, as the changes would be little more hefty.

Signed-off-by: Andrew Harding <aharding@vmware.com>