Removing plugin related materials from DataStore
Lots of artifacts around DataStore pluggability remain. This is the
first step to removing some of those artifacts. Two reasons for doing
this now:

1. The plugingen tool is used to generate the DataStore interface from
the gRPC service definition. The plugingen tool will be removed very
soon as part of the catalog refactor and plugin SDK introduction (#2153)

2. Since all of the scaffolding is in place for the datastore, its easy
for test code to load it as a plugin, which is DIFFERENT than SPIRE
server loads it. This can lead to an divergence in expected behavior.

Cleaning the datastore interface up completely is out of scope for this
PR. This PR does not take the change as far as it could, for sake of
review simplicity. For example, it maintains the request/response
parameters even when those requests/response should eventually be
removed in most cases where we have few or no fields in those structs.
The SQL plugin also still takes HCL string configuration as input
instead of a already populated struct.

Most of the changes are scoped to the de-protobuf'ing of the
request/response structs. The creation and configuration path was also
tweaked to be not be reliant on the plugin loading paradigms (e.g.
SetLogger, Configure RPC). The logger was also changed to the logrus
logger the rest of SPIRE uses instead of the hclog for plugins.

Signed-off-by: Andrew Harding <aharding@vmware.com>