Restructure v0 plugin interface layout
Aligns the directories with the plugin-sdk.

Most of the code changes outside of the import path updates are limited
to spots where we missed using the versioned import when introducing the
facades.

While I was in the position of regenerating all of the gRPC code, I
pulled in version 1.1.0 of the protoc-gen-go-grpc tool.

Signed-off-by: Andrew Harding <aharding@vmware.com>