Fix race in agent node attestation tests
The fake server stream allows tests to configure an attestation failure.
When set, the fake server stream currently fails the attestation call
before the `Recv` for the client attestation data. Left to the whims of
timing, this causes the client code to observe the failure either on
their own `Recv` of the response (the test assumes this) or earlier when
the client tries to `Send` (observed as an EOF, since the stream has
already been closed). The latter causes a false-negative test failure.

This change fixes the raciness by moving the code that fails attestation
until after the attestation data has been received by the client so that
the test consistently observes the failure on the `Recv`.

Signed-off-by: Andrew Harding <aharding@vmware.com>