Fix premature stream closure handling in v0 nodeattestor shim
There is a unit-test that asserts that the shim properly identifies and
handles instances where a plugin closes the stream prematurely and
therefore cannot handle a challenge issued by the server.

The test harness facilitates this by having the plugin close the stream
right after sending the initial attestation data. The shim would detect
the condition by checking for io.EOF when sending the challenge on the
stream. However, if timing is just right, the shim may not observe the
stream closure until attempting to receive the challenge response.

This code fixes the shim to also detect the EOF condition on the receive
path. Note that nothing was "broken" in that attestation would still
fail, but the error message isn't as helpful without the fix.

Signed-off-by: Andrew Harding <aharding@vmware.com>