Fix racy manager test
The racy test was one which verified that the manager properly notified
notifiers of the bundle loaded event when the manager is first run.
Since run blocks until the caller context is canceled, the fake notifier
would cancel the context when it successfully received the notification.
The test assumed that it would observe the cancelation in the error
status reported from `Run`. However, this was not always so due to
timing.

This change fixes the racy test by accepting either no error or the
cancelation error coming out of the notifier.

Signed-off-by: Andrew Harding <aharding@vmware.com>