Remove Noop NodeResolver plugin
The noop noderesolver dates back to a time before node resolvers were
made optional and paired with node attestors. As node resolvers are no
longer required, it is an unneccessary plugin and its presence in the
codebase and in documentation is a source of confusion.

This change rips the noop noderesolver plugin out. Considering the noop
plugin has been present in the sample documentation for quite some time,
it is possible that existing deployments have it configured. As such,
the noop configuration is detected and a warning message logged.

Signed-off-by: Andrew Harding <aharding@vmware.com>