Do not expire Prometheus metrics by default
The `armon/go-metrics` package by default drops emission of Prometheus metrics
after 60 seconds. This behavior is undesirable as the default because it
breaks dashboard and alert queries. SPIRE metrics in general are meant
to be continually emitted and not unregistered, so switching the default
behavior to not drop metrics after 60 seconds.

Signed-off-by: Ryan Turner <turner@uber.com>