Fix test race introduced by recent refactor
PR #2220 broke the test by swapping out time in the `now` variable
(which is the expiration time) with `time.Now()`. When enough time
elapses that the second boundary is crossed, the token IS pruned and the
test fails.

Also fixed the test to fail gracefully if this happens again instead of
panicking.

Signed-off-by: Andrew Harding <aharding@vmware.com>