Fix node selector query split-response
The DataStore query that returns the list of agent node selectors is
currently sorted by row ID, which under certain circumstances can cause
it to split the selectors for a given agent ID into more than one entry
in the aggregated results. The only caller is the full entry cache,
which does not handle this well, which can cause agents to not be
associated with a node alias whose selectors end up spanning these
splits.

This change fixes the query to order by SPIFFE ID instead of row ID so
that the selectors for a given agent are grouped together and returned
in a single entry in the aggregated results.

It also adds a defensive measure in the full cache layer to also
implement robust handling of the results so that a bug like this does
not impact cache correctness in the future. This has a small temporary
memory cost while aggregating the results, but it should be minimal
compared to the cost of the full cache itself.

Signed-off-by: Andrew Harding <aharding@vmware.com>