Adding datastore health check.
Health checks for datastore are added along with tests in the catalog.

* pkg/common/health/health.go
Removed `LiveState` and `ReadyState` field from the Checker interface as
they weren't used anywhere except in the handlers in the file itself.

* pkg/server/catalog/catalog.go
Added/registered the datastore health check.

* pkg/server/plugin/datastore/datastore_health.go
The implementation of the datastore health check which is done by
checking the datastore's ability to list bundles for both liveness and
readiness checks.

* pkg/server/plugin/datastore/sql/models.go
* pkg/server/plugin/datastore/sql/sql.go
Fixes a few minor errors.

* pkg/server/server.go
Added the datastore health check via catalog.

* test/fakes/fakehealthchecker/checker.go
Removed the `LiveState` and `ReadyState` methods which are removed from
the Checker interface in `health.go`.

Signed-off-by: Sachin Kumar Singh <sachinkumarsingh092@gmail.com>