Backport fix node selector query split-response to 0.12 (#2242)
The DataStore query that returns the list of agent node selectors is currently sorted by row ID, which under certain circumstances can cause it to split the selectors for a given agent ID into more than one entry in the aggregated results. The only caller is the full entry cache, which does not handle this well, which can cause agents to not be associated with a node alias whose selectors end up spanning these splits.

This change fixes the query to order by SPIFFE ID instead of row ID so that the selectors for a given agent are grouped together and returned in a single entry in the aggregated results.

It also aggregates the results differently such that the order does not matter for correctness so that the code is resilient against a similar bug in the future.

Co-Authored-By: Andrew Harding <aharding@vmware.com>
Signed-off-by: Jonathan Oddy <jonathan.oddy@transferwise.com>