Use ca_key_type for the server SVID
Before this the server SVID used to be hardcoded use
EC curve P384. This is different from the default key
type and can be different from what is configured for
other SVIDs.

Since this SVID is visible to other systems, e.g. when
federating, make it use the configure ca key type.

Fixes: https://github.com/spiffe/spire/issues/2127

Signed-off-by: Sorin Dumitru <sdumitru@bloomberg.net>