Remove RSA 1024 support from key manager facades
SPIRE does not deal in 1024-bit RSA keys (and shouldn't) but the server
keymanager interface has the key type defined and the built in plugins
support it.

This change removes support for 1024-bit RSA keys from the key manager
facades and built-ins. Old plugins that still implement support for
these keys are unaffected by this change. The v1 key manager interface
(to be released officially with 1.0) no longer defines this key type so
new plugins going forward won't pick up support for it needlessly.

Signed-off-by: Andrew Harding <aharding@vmware.com>