Fix join_token plugin error message (#2299)
The agent handles the case where the join token is set using special
cased code. The special code is only activated when the join token has
been provided via CLI flag or HCL configuration, whether or not the
join_token node attestor has been configured. If the join token is not
set, but the join_token node attestor is configured, then the special
case code will not be activated and this plugin will end up being
invoked.

This change update the error message returned by the plugin to inform
operators that the join token was not provided.

Signed-off-by: Andrew Harding <aharding@vmware.com>