Fix race condition in certmanager UpstreamAuthority unit test
The `Test_MintX509CA` unit test spawns another goroutine that attempts
to update the status of a Kubernetes CertificateRequest object during
execution of the `MintX509CA` streaming API handler. The implementation
of the test was prone to race conditions as it was relying on timing
factors between the two goroutines in the unit test and the API handler.
If the main goroutine of the unit test completes before the spawned
goroutine finishes executing, the test will fail because the
CertificateRequest object is not yet in the expected state.

Eliminate the usage of `time.Sleep()` as an attempt to coordinate
between the test and the API handler and replace it with hooks that send
events over channels. The test coordinates the sequence of events it
needs to perform using these channels.

Signed-off-by: Ryan Turner <turner@uber.com>