Remove debug logging emitted from CA health check
The CA health check operates by having the CA subsystem sign an SVID.
However, this produces a debug log that muddies the logs and isn't
an operator concern.

This change introduces a way for health checks to pass a context to
subsystems that is marked as a health check context. The subsystem can
then query the context and make any necessary adjustments as necessary.

It isn't quite what I'd like and is a short-term fix. Ideally:

- CheckHealth is passed a context by the health subsystem that has
already been marked, so health checkers don't have to remember to do
this. I'm not doing this now since it would require a shakeup to the
health package that isn't quite appropriate given the proximity to the
1.0 release.
- Subsystems query for loggers and metrics interfaces from the context
instead of getting their own logger. This allows us to set noop
implementations on the context before passing it in, which removes the
need to mix in health concerns (e.g. checking if the context is a health
check before logging) into each subsystem.

Signed-off-by: Andrew Harding <aharding@vmware.com>