Fix race in server nodeattestor facade tests (#2380)
There are test cases that ensure that the facade properly wraps errors
that are returned by the plugin before the first receive has been
issued. However, the way that is observed on the gRPC client side is not
deterministic.  It may show up as an EOF, or as the actual error,
depending wholly on the timing of the Send.

This change updates the tests to not check for specific gRPC status
codes or messages on the return, only that a properly wrapped error was
returned.

Signed-off-by: Andrew Harding <aharding@vmware.com>