Back out AuditLogEnabled server configurable
The audit log feature is still under development and should not be
accessible to users until it is completed. This change is motivated by
the desire to keep the main branch of SPIRE release-able, such that
maintainers do not need to manipulate a set of cherry-picked commits
to craft patch releases.

Signed-off-by: Ryan Turner <turner@uber.com>