Limit cipher suites used by the bundle endpoint
We need to constrain the cipher suites available on the bundle endpoint
when ACME is enabled since the KeyManager manages the keys and there are
limitations on some KeyManagers as to what signature algorithms are
supported (see issue #2302).

This change updates the TLS config to only support cipher suites that
use SHA256 for the ECP256 and RSA2048 keys generated by the autocert
package. This aligns to what is supported by AWS KMS.

Signed-off-by: Andrew Harding <aharding@vmware.com>