Update TTL warning messages to detect all possible conditions
Turns out there are actually three modes of surprise with regards to TTL
configuration:
* SVID TTL is too high for the CA TTL
* SVID TTL is too high period (there is actually a 7 day cap)
* SVID TTL is too high period AND the CA TTL is too low

This commit adds logic to detect and warn on each of these conditions.
It also moves the exact timing and compatibility logic in to the `ca`
package since it is the one responsible for the rotation algorithms.

Note that the CA currently implements an "activation threshold cap".
This means that if the activation threshold would normally be longer
than 7 days (given the normal lifetime of the CA), then activate 7 days
in advance rather than some greater period of time.

The intent of this logic is to minimize overhead when the ca_ttl is
long. For example, if the ca_ttl is set to two years we probably don't
want to activate the new CA four months in advance. The consequence of
this is that we can never guarantee an SVID TTL longer than 7 days.

This rotation algorithm could definitely use some improvement, as is
evidenced by the difficulty in capturing its behavior here.

Signed-off-by: Evan Gilman <egilman@vmware.com>