Improve log messages about configured ttl values (#2284)
* Fix log messages about configured ttl values
* Keep CA rotation timing logic inside the CA package

This commit builds upon @hiyosi's previous by extracting logic related
to the timing of CA rotation, and instead exposing methods from the CA
package to obtain the information necessary. I additionatlly un-exported
some methods that were previously exported for consumption in the CLI
which we no longer need.

We appear to have been logging two messages if the CA TTL and SVID TTL
were out of sync rather than one. I removed one of the messages.

* Update TTL warning messages to detect all possible conditions

Turns out there are actually three modes of surprise with regards to TTL
configuration:
* SVID TTL is too high for the CA TTL
* SVID TTL is too high period (there is actually a 7 day cap)
* SVID TTL is too high period AND the CA TTL is too low

This commit adds logic to detect and warn on each of these conditions.
It also moves the exact timing and compatibility logic in to the `ca`
package since it is the one responsible for the rotation algorithms.

Note that the CA currently implements an "activation threshold cap".
This means that if the activation threshold would normally be longer
than 7 days (given the normal lifetime of the CA), then activate 7 days
in advance rather than some greater period of time.

The intent of this logic is to minimize overhead when the ca_ttl is
long. For example, if the ca_ttl is set to two years we probably don't
want to activate the new CA four months in advance. The consequence of
this is that we can never guarantee an SVID TTL longer than 7 days.

This rotation algorithm could definitely use some improvement, as is
evidenced by the difficulty in capturing its behavior here.

* Please the linter by switching to switch
* Fix some typos in spire-server run cli

Signed-off-by: Tomoya Usami <tousami@zlab.co.jp>
Signed-off-by: Evan Gilman <egilman@vmware.com>
Co-authored-by: Evan Gilman <egilman@vmware.com>