Fix kubernetes-based integration tests
Kind 0.10 started failing on linux for some reason. I dug for a while
but couldn't figure it out. The node would not enter a ready state. This
started failing at roughly the same time for me, both locally and on
CI/CD.

In any case, 0.11.1 works fine, so I've migrated to integration tests to
use it. I've pinned the node version to Kubernetes v1.20 to maintain the
same version that was being used with Kind 0.10.

I've also updated the tooling to use a pinned version of kubectl instead
of grabbing the latest.

Since this is contributing to a CI/CD failure that we need sorted
somewhat quickly, I didn't take the time to refactor some of these
Kubernetes cluster setup bits that have been copy/pasted into multiple
integration tests now, but that would be an excellent contribution at a
later date.

Signed-off-by: Andrew Harding <aharding@vmware.com>